+layers first draft

+++ deckard/layers/parse.py
@@ -23,11 +23,11 @@ logger = logging.getLogger(__name__)
 )
 def hydra_parser(cfg: DictConfig, **kwargs):
     params = OmegaConf.to_object(cfg)
-    params = parse(params, **kwargs)
+    params = generate_paths_from_params(params, **kwargs)
     print(yaml.dump(params, indent=4))
     logger.info("Successfully parsed the hydra config and saved it to params.yaml")
 
-def parse(params:dict, queue_path="queue", default = "params.yaml", filename = "default.yaml") -> dict:
+def generate_paths_from_params(params:dict, queue_path=None, default = "params.yaml", filename = "default.yaml") -> dict:
     """
     Parse the hydra config and save it to a yaml file in both the queue_path and the current working directory with the name "params.yaml".
     :param params: The hydra config as a dictionary
@@ -74,20 +74,55 @@ def parse(params:dict, queue_path="queue", default = "params.yaml", filename = "
     params["files"]["path"] = str(my_hash(params)) if params["files"]["path"] is None else params["files"]["path"]
     params = DictConfig(params)
     params = OmegaConf.to_container(params, resolve=True)
-    if Path(default).exists():
-        Path(os.getcwd(), default).unlink()
-    filename = Path(queue_path, my_hash(params) + ".yaml") if filename is None else Path(queue_path, filename)
-    filename.parent.mkdir(parents=True, exist_ok=True)
-    # Write the params to the queue
-    with open(Path(filename), "w") as f:
-        yaml.dump(params, f)
-    # Write the params to the default
-    with open(Path(default), "w") as f:
-        yaml.dump(params, f)
-    logger.info(f"Wrote params to {filename} and {default}")
     return params
 
+def dump_params_to_queue(params:dict, queue_path=None, filename = None) -> None:
+    # Write the params to the queue
+    if queue_path is not None:
+        filename = Path(queue_path, my_hash(params) + ".yaml") if filename is None else Path(queue_path, filename)
+        filename.parent.mkdir(parents=True, exist_ok=True)
+        with open(Path(filename), "w") as f:
+            yaml.dump(params, f)
+    logger.info(f"Wrote params to {filename}")
+    return None     
+            
+def dump_params_to_default(params:dict, default:str = "params.yaml") -> None:
+    # Write the params to the default
+    if default is not None:
+        if Path(default).exists():
+            Path(os.getcwd(), default).unlink()
+        with open(Path(default), "w") as f:
+            yaml.dump(params, f)
+    logger.info(f"Wrote params to {default}")
+    return None
 
+
+
+def generate_stage_from_params(stage, params:dict, filename="dvc.yaml", output_folder="reports",) -> None:
+    """
+    Generate a stage from the params
+    :param stage: The stage to generate
+    :param params: The params to generate the stage from
+    :param filename: The name of the file to save the stage to
+    :return: None
+    """
+    # Load the dvc.yaml file
+    with open(Path(filename), "w") as f:
+        pipe = yaml.load(f, Loader=yaml.FullLoader)['stages'][stage]
+    # If there is a dictionary inside of the params, we need to flatten it
+    flattened = flatten_dict(params)
+    # Replace the params in the dvc.yaml (denoted by "${ }") file with the flattened params
+    for key, value in flattened.items():
+        for ki, vi in pipe.items():
+            if isinstance(vi, str):
+                pipe[ki] = vi.replace("${" + key + "}", str(value))
+        
+    
+def flatten_dict():
+    pass
+
+def generate_grid():
+    pass
 if "__main__" == __name__:
     _ = hydra_parser()
     assert Path("params.yaml").exists(), \
